backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
openai==1.3.5
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2
pytest==7.4.3
httpx==0.25.2
bcrypt==4.0.1
alembic==1.12.1 